Take care to avoid integer overflow when doing the initial page cache allocation with an excessively large cache_size setting.
FossilOrigin-Name: 4d66ac98deaa85218be7ff0eb254f78b96d8e8d4

+++ src/pcache1.c
@@ -279,7 +279,7 @@ static int pcache1InitBulk(PCache1 *pCache){
     szBulk = -1024 * (i64)pcache1.nInitPage;
   }
   if( szBulk > pCache->szAlloc*(i64)pCache->nMax ){
-    szBulk = pCache->szAlloc*pCache->nMax;
+    szBulk = pCache->szAlloc*(i64)pCache->nMax;
   }
   zBulk = pCache->pBulk = sqlite3Malloc( szBulk );
   sqlite3EndBenignMalloc();
